refactor: merge DRIVE_SYNC_ERROR_CODES into api.js
Move the constant from standalone driveSyncErrorCodes.js into
scripts/config/api.js alongside other API-related constants,
reducing file count while preserving the same single-source-of-truth.
